+++ packages/cmr-client/tests/test-CMR.js
@@ -5,7 +5,7 @@ const nock = require('nock');
 const some = require('lodash/some');
 
 const awsServices = require('@cumulus/aws-client/services');
-const { ValidationError } = require('@cumulus/errors');
+const { CMRInternalError } = require('@cumulus/errors');
 
 const { CMR } = require('../CMR');
 
@@ -165,6 +165,32 @@ test('getReadHeaders returns clientId and token for launchpad', (t) => {
   t.is(headers.Authorization, '12345');
 });
 
+test.serial.only('ingestUMMGranule() returns CMRInternalError when CMR is down', async (t) => {
+  const cmrSearch = new CMR({ provider: 'my-provider', token: 'abc', clientId: 'client' });
+
+  const ummgMetadata = { GranuleUR: 'asdf' };
+
+  const internalError = {
+    errors: [
+      {
+        errors: ['Internal error'],
+      },
+    ],
+  };
+
+  process.env.CMR_ENVIRONMENT = 'SIT';
+
+  nock('https://cmr.sit.earthdata.nasa.gov')
+    .put(`/ingest/providers/${cmrSearch.provider}/granules/${ummgMetadata.GranuleUR}`)
+    .times(3)
+    .reply(503, internalError);
+
+  await t.throwsAsync(
+    () => cmrSearch.ingestUMMGranule(ummgMetadata),
+    { instanceOf: CMRInternalError }
+  );
+});
+
 test.serial('ingestUMMGranule() throws an exception if the input fails validation', async (t) => {
   const cmrSearch = new CMR({ provider: 'my-provider', token: 'abc', clientId: 'client' });
 
@@ -182,12 +208,15 @@ test.serial('ingestUMMGranule() throws an exception if the input fails validatio
   process.env.CMR_ENVIRONMENT = 'SIT';
 
   nock('https://cmr.sit.earthdata.nasa.gov')
-    .post(`/ingest/providers/${cmrSearch.provider}/validate/granule/${ummgMetadata.GranuleUR}`)
+    .put(`/ingest/providers/${cmrSearch.provider}/granules/${ummgMetadata.GranuleUR}`)
     .reply(422, ummValidationError);
 
   await t.throwsAsync(
     () => cmrSearch.ingestUMMGranule(ummgMetadata),
-    { instanceOf: ValidationError }
+    {
+      name: 'Error',
+      message: 'Failed to ingest, statusCode: 422, statusMessage: Unprocessable Entity, CMR error message: [{"path":["Temporal"],"errors":["oh snap"]}]',
+    }
   );
 });